added: MIOpenDriver for PadConstant

+++ include/miopen/miopen.h
@@ -6602,7 +6602,7 @@ MIOPEN_EXPORT miopenStatus_t miopenPadConstantFwd(miopenHandle_t handle,
                                                   miopenTensorDescriptor_t yDesc,
                                                   const void* x,
                                                   void* y,
-                                                  const int* padding,
+                                                  const size_t* padding,
                                                   float value);
 
 #endif // MIOPEN_BETA_API
