+++ src/include/miopen/constant_pad.hpp
@@ -38,7 +38,7 @@ miopenStatus_t PadConstantForward(Handle& handle,
                                    const TensorDescriptor& yDesc,
                                    ConstData_t x,
                                    Data_t y,
-                                   const int* padding,
+                                   const size_t* padding,
                                    float value);
 } // namespace miopen
 #endif
